changed auth0 code and put in server.js
